Change in personal process.

+++ regression_test/run_unit_regressions.sh
@@ -5,17 +5,23 @@ COLLECTIONS=( neossat gem omm vlass draost cgps cfht dao phangs )
 
 # provide a collection name as a parameter to 'run just one' set of
 # unit tests
-if [[ $# -eq 1 ]]
-then
-  test_set=( "${1}" )
-else
-  test_set=( ${COLLECTIONS[@]} )
+# if [[ $# -eq 1 ]]
+# then
+#   test_set=( "${1}" )
+# else
+#   test_set=( ${COLLECTIONS[@]} )
+# fi
+
+if [[ $# -eq 4 ]]; then
+  echo "Usage: ${0} <collection> <git repo name> <git branch name>"
+  exit 1
 fi
 
-cadc_repo="opencadc"
-omc_repo="opencadc"
-# opencadc_branch="CADC-10281"
-opencadc_branch="master"
+test_set=( "${1}" )
+git_repo="${2:-opencadc}"
+git_branch="${3:-master}"
+
+echo "Executing regression test for ${collection} from ${git_repo}/${git_branch}"
 
 mkdir -p executions || exit $?
 cd executions || exit $?
@@ -35,8 +41,13 @@ do
     echo "Using data mount ${data_mount}"
   fi
 
-  echo ":::retrieve Dockerfile for ${collection} from https://raw.github.com/${omc_repo}/${collection}2caom2/master/Dockerfile"
-  curl -L https://raw.github.com/${omc_repo}/${collection}2caom2/master/Dockerfile -o Dockerfile || exit $?
+  if [[ ${git_branch} == "master" ]]; then
+    echo ":::retrieve Dockerfile for ${collection} from https://raw.github.com/${git_repo}/${collection}2caom2/master/Dockerfile"
+    curl -L https://raw.github.com/${git_repo}/${collection}2caom2/master/Dockerfile -o Dockerfile || exit $?
+  else
+    echo ":::retrieve Dockerfile for ${collection} from https://raw.github.com/${git_repo}/${collection}2caom2/raw/${git_branch}/Dockerfile"
+    curl -L https://github.com/${git_repo}/${collection}2caom2/raw/${git_branch}/Dockerfile -o Dockerfile || exit $?
+  fi
   
   echo ":::docker build for ${collection}"
   sudo rm *xml 
@@ -48,7 +59,7 @@ do
       sudo rm test_files/success/*.fits* 
       sudo rm test_files/failure/*.fits* 
   fi
-  sudo docker build -f Dockerfile --build-arg OPENCADC_BRANCH=${opencadc_branch} --build-arg OPENCADC_REPO=${cadc_repo} --build-arg CAOM2_BRANCH=${opencadc_branch} --build-arg CAOM2_REPO=${cadc_repo} -t ${collection} ./ || exit $?
+  sudo docker build -f Dockerfile --build-arg OPENCADC_BRANCH=${git_branch} --build-arg OPENCADC_REPO=${git_repo} -t ${collection} ./ || exit $?
 
   echo "::: build ingest config"
   cp $HOME/.ssl/cadcproxy.pem . || exit $?
@@ -70,14 +81,17 @@ do
   echo ""
   echo ""
   echo ""
-  echo ":::docker run ${collection}_run SCRAPE MODIFY"
-  sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} ${collection}_run || exit $?
-  cp ../../does_collection_clean_up.py . || exit $?
-  cp ../../compare_run.py . || exit $?
-  sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} python compare_run.py ${collection} || exit $?
-  echo ""
-  echo ""
-  echo ""
+  if [[ ${collection} != "gem" ]]; then
+     # can't run gem2caom2 with SCRAPE
+     echo ":::docker run ${collection}_run SCRAPE MODIFY"
+     sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} ${collection}_run || exit $?
+     cp ../../does_collection_clean_up.py . || exit $?
+     cp ../../compare_run.py . || exit $?
+     sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} python compare_run.py ${collection} || exit $?
+     echo ""
+     echo ""
+     echo ""
+  fi
   echo ":::docker run ${collection}_run INGEST ${PWD}"
   sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} python build_ingest_config.py ${collection} || exit $?
   sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} ${collection}_run || exit $?
@@ -89,7 +103,12 @@ do
     cp ../../does_collection_clean_up.py . || exit $?
     cp ../../build_state.py . || exit $?
     sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} python build_state.py ${collection} || exit $?
-    sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} ${collection}_run_incremental || exit $?
+
+      if [[ ${collection} = "gem" ]]; then
+        sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} ${collection}_run_incremental || exit $?
+      else
+        sudo docker run --rm -v ${PWD}:/usr/src/app/ ${data_mount} ${collection} ${collection}_run_state || exit $?
+      fi
   fi
   cd .. || exit $?
   echo ":::${collection} Success at: $(date)" >> ../success_log.txt
